exception handling added, enums separated, dependencies version changed to get swagger work
