+++ conveyor/src/main/java/com/example/conveyor/exception/ScoringException.java
@@ -1,0 +1,7 @@
+package com.example.conveyor.exception;
+
+public class ScoringException extends RuntimeException{
+    public ScoringException(String message) {
+        super(message);
+    }
+}
