+++ conveyor/src/main/java/com/example/conveyor/model/EmploymentStatus.java
@@ -1,0 +1,8 @@
+package com.example.conveyor.model;
+
+public enum EmploymentStatus {
+    EMPLOYED,
+    UNEMPLOYED,
+    SELF_EMPLOYED,
+    BUSINESS_OWNER
+}
