+++ conveyor/src/main/java/com/example/conveyor/model/Gender.java
@@ -1,0 +1,7 @@
+package com.example.conveyor.model;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+    NON_BINARY
+}
