+++ conveyor/src/main/java/com/example/conveyor/model/MaritalStatus.java
@@ -1,0 +1,7 @@
+package com.example.conveyor.model;
+
+public enum MaritalStatus {
+    SINGLE,
+    MARRIED,
+    DIVORCED
+}
