+++ conveyor/src/main/java/com/example/conveyor/model/Position.java
@@ -1,0 +1,7 @@
+package com.example.conveyor.model;
+
+public enum Position{
+    ORDINARY_WORKER,
+    MIDDLE_MANAGER,
+    TOP_MANAGER
+}
