+++ conveyor/src/main/resources/open-api/dtos.yml
@@ -1,0 +1,89 @@
+openapi: 3.0.3
+info:
+  title: Кредитный конвейер - OpenAPI 3.0
+  description: Микросервис конвейер
+  version: 1.0.11
+servers:
+  - url: https://localhost:8080/
+paths: { }
+
+components:
+  schemas:
+    CreditDTO:
+      type: object
+      description: Параметры кредита после скоринга
+      properties:
+        amount:
+          type: number
+          description: Сумма кредита
+          format: double
+          example: 200000.0
+        term:
+          type: integer
+          description: Продолжительность выплаты кредита в месяцах
+          format: int32
+          example: 36
+        monthlyPayment:
+          type: number
+          description: Ежемесячный платёж
+          format: double
+          example: 2000.0
+        rate:
+          type: number
+          description: Годовая ставка кредита в процентах
+          format: double
+          example: 7.6
+        psk:
+          type: number
+          description: Полная стоимость кредита в процентах
+          format: double
+          example: 7.6
+        isInsuranceEnabled:
+          type: boolean
+          description: Страховка включена
+        isSalaryClient:
+          type: boolean
+          description: Зарплатный клиент
+        paymentSchedule:
+          type: array
+          items:
+            $ref: '#/components/schemas/PaymentScheduleElement'
+
+    PaymentScheduleElement:
+      type: object
+      description: Регулярный платёж
+      properties:
+        number:
+          type: integer
+          description: Порядковый номер выплаты
+          format: int32
+          example: 1
+        date:
+          type: string
+          description: Дата выплаты
+          format: date
+          example: 2023-08-01
+        totalPayment:
+          type: number
+          format: double
+          example: 7000.0
+          description: Сумма выплаты
+        interestPayment:
+          type: number
+          description: Сумма выплаты процентов
+          format: double
+          example: 3000.0
+        debtPayment:
+          type: number
+          description: Сумма выплаты основного долга
+          format: double
+          example: 4000.0
+        remainingDebt:
+          type: number
+          description: Остаток долга
+          format: double
+          example: 287421.75
+
+
+
+
